# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MissionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MissionResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "MissionListResponse",
        "type": "object"
      },
      "MissionOutcome": {
        "enum": [
          "SUCCESS",
          "PARTIAL_SUCCESS",
          "HUMAN_REVIEW_REQUIRED",
          "BLOCKED",
          "FAILED",
          "CANCELLED",
          "EXPIRED",
          "REJECTED",
          "INTERRUPTED"
        ],
        "title": "MissionOutcome",
        "type": "string"
      },
      "MissionPhase": {
        "enum": [
          "QUEUED",
          "OFFERED",
          "ACCEPTED",
          "NAVIGATING",
          "WORKING",
          "RETURNING",
          "TERMINAL"
        ],
        "title": "MissionPhase",
        "type": "string"
      },
      "MissionRequest": {
        "additionalProperties": false,
        "properties": {
          "idempotency_key": {
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "requested_by": {
            "minLength": 1,
            "title": "Requested By",
            "type": "string"
          },
          "seat_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seat Id"
          },
          "target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MissionTargetModel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "priority",
          "requested_by",
          "idempotency_key"
        ],
        "title": "MissionRequest",
        "type": "object"
      },
      "MissionResponse": {
        "properties": {
          "after_observation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Observation"
          },
          "before_observation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Observation"
          },
          "cancel_requested": {
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "mission_id": {
            "title": "Mission Id",
            "type": "string"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MissionOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "phase": {
            "$ref": "#/components/schemas/MissionPhase"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "requested_by": {
            "title": "Requested By",
            "type": "string"
          },
          "seat_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seat Id"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "target": {
            "$ref": "#/components/schemas/MissionTargetModel"
          }
        },
        "required": [
          "mission_id",
          "target",
          "seat_id",
          "priority",
          "requested_by",
          "idempotency_key",
          "created_at",
          "phase",
          "outcome",
          "message",
          "sequence",
          "cancel_requested",
          "before_observation",
          "after_observation"
        ],
        "title": "MissionResponse",
        "type": "object"
      },
      "MissionTargetModel": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/TargetKind"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "reference_id": {
            "minLength": 1,
            "title": "Reference Id",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "reference_id"
        ],
        "title": "MissionTargetModel",
        "type": "object"
      },
...
      "Priority": {
        "enum": [
          "NORMAL",
          "HIGH"
        ],
        "title": "Priority",
        "type": "string"
      },
...
      "TargetKind": {
        "enum": [
          "SEAT",
          "ZONE",
          "POINT"
        ],
        "title": "TargetKind",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Cleany Control Plane",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/missions": {
      "get": {
        "operationId": "list_missions_api_missions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Missions"
      },
      "post": {
        "operationId": "create_mission_api_missions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Mission"
      }
    },
    "/api/missions/{mission_id}/cancel": {
      "post": {
        "operationId": "cancel_mission_api_missions__mission_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Mission"
      }
    },
...
  }
}